+++ src/main/java/com/obiscr/chatgpt/icons/ChatGPTIcons.java
@@ -1,0 +1,18 @@
+package com.obiscr.chatgpt.icons;
+
+import com.intellij.openapi.util.IconLoader;
+
+import javax.swing.*;
+
+/**
+ * @author wuzi
+ */
+public class ChatGPTIcons {
+
+    public static final Icon ME = IconLoader.getIcon("/images/me.png",ChatGPTIcons.class);
+
+    public static final Icon AI = IconLoader.getIcon("/images/ai.png",ChatGPTIcons.class);
+
+    public static final Icon OPEN_AI = IconLoader.getIcon("/images/openai.png",ChatGPTIcons.class);
+
+}