+++ src/main/java/com/obiscr/chatgpt/ui/action/SettingAction.java
@@ -4,7 +4,7 @@ import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.options.ShowSettingsUtil;
 import com.intellij.openapi.project.DumbAwareAction;
-import com.obiscr.chatgpt.settings.ChatGPTSettingsPanel;
+import com.obiscr.chatgpt.settings.OpenAISettingsPanel;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NotNull;
 
@@ -19,6 +19,6 @@ public class SettingAction extends DumbAwareAction {
 
   @Override
   public void actionPerformed(@NotNull AnActionEvent e) {
-    ShowSettingsUtil.getInstance().showSettingsDialog(e.getProject(), ChatGPTSettingsPanel.class);
+    ShowSettingsUtil.getInstance().showSettingsDialog(e.getProject(), OpenAISettingsPanel.class);
   }
 }