+++ src/main/java/com/obiscr/chatgpt/ui/action/editor/AbstractEditorAction.java
@@ -4,18 +4,23 @@ import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.editor.Editor;
-import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.NlsActions;
+import com.obiscr.chatgpt.core.SendAction;
+import com.obiscr.chatgpt.message.ChatGPTBundle;
+import com.obiscr.chatgpt.ui.MainPanel;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.function.Supplier;
+
+import static com.obiscr.chatgpt.MyToolWindowFactory.ACTIVE_CONTENT;
 
 /**
  * @author Wuzi
  */
 public abstract class AbstractEditorAction extends AnAction {
 
-    protected String selectedText = null;
+    protected String text = "";
+    protected String key = "";
 
     public AbstractEditorAction(@NotNull Supplier<@NlsActions.ActionText String> dynamicText) {
         super(dynamicText);
@@ -23,7 +28,13 @@ public abstract class AbstractEditorAction extends AnAction {
 
     @Override
     public void actionPerformed(@NotNull AnActionEvent e) {
-
+        Editor editor = e.getData(CommonDataKeys.EDITOR);
+        assert editor != null;
+        String selectedText = editor.getSelectionModel().getSelectedText();
+        text = ChatGPTBundle.message(key, selectedText);
+        SendAction sendAction = e.getProject().getService(SendAction.class);
+        Object mainPanel = e.getProject().getUserData(ACTIVE_CONTENT);
+        sendAction.doActionPerformed((MainPanel) mainPanel, text);
     }
 
     @Override