+++ src/main/java/com/obiscr/chatgpt/util/StringUtil.java
@@ -1,8 +1,7 @@
 package com.obiscr.chatgpt.util;
 
-import com.alibaba.fastjson2.JSON;
-import com.alibaba.fastjson2.JSONArray;
-import com.alibaba.fastjson2.JSONObject;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * @author Wuzi
@@ -21,11 +20,11 @@ public class StringUtil extends com.intellij.openapi.util.text.StringUtil {
         return true;
     }
 
-    public static String appendMe(String question) {
-        return "> ![](https://intellij-icons.jetbrains.design/icons/AllIcons/general/user.svg) 我\n\n" + question + "\n\n";
-    }
+    public static @NotNull String stripHtml(@NotNull String html, @Nullable String breaks) {
+        if (breaks != null) {
+            html = html.replaceAll("<br/?>", breaks);
+        }
 
-    public static String appendQuestion(String response) {
-        return "\n\n> ![](https://intellij-icons.jetbrains.design/icons/AllIcons/general/balloonInformation.svg) ChatGPT\n\n" + response + "\n\n";
+        return html.replaceAll("<(.|\n)*?>", "");
     }
 }